- Few changes to my scene. - Had to edit Project and Build settings to build an Android version.
Would also just like to point out, please don't edit code that other people's scenes are relying on before asking permission. Or if you need to, copy and paste the code and make your own version of that until we figure out an easy way to combine our efforts without breaking 3 entire systems of a demo that was due today. Luckily they didn't really take a look at any of our new changes. But next time we might not be so lucky.

Also ask questions if you don't know how to make a new enemy or prefabs. I saw some projectile prefabs using old scripts - from before we had a health system. We don't want to use Destroy On Impact script anymore, unless it is a weapon that is 100% of the time a one-hit kill.